requirements.txt
Django>=4.0,<5.0
djangorestframework
requests
Pillow
androguard
dj-rest-auth[with_social]
django-rest-auth
oss2
